refactor: extract execution-contract types into homeboy-execution-contract
The execution_contract module was almost entirely pure: the
ExecutionContract/ArtifactUriContract/LabOffloadExecutionContract/
ApplyChangeContract types, the EXECUTION_CONTRACT const, the pure
classification (is_remote_runner_artifact_path) and base64 token helpers
(runner_artifact_store_token, artifact_store_locator_from_runner_artifact_id).
The ONLY thing forcing it to stay in core was is_reportable_artifact_evidence_path,
whose final branch probes the filesystem via std::fs::metadata.

Now that ArtifactReference is a leaf (homeboy-artifact-ref-contract, #8997),
ArtifactReference::parse no longer drags in observation, so the pure surface
can move DOWN into a new leaf crate homeboy-execution-contract (deps:
artifact-ref-contract + engine-primitives + base64). The two fs-touching
functions (is_reportable_artifact_evidence_path / reportable_artifact_evidence_path)
stay in core as free functions.

Core re-exports the contract as execution_contract, so all existing paths
(core-internal + agents + lab-runner + cli) resolve unchanged.

This completes the execution_contract extraction the campaign was building
toward -- the last edge was severed by the artifact-ref relocation.
